+++ data/resources/CHANGELOG.md
@@ -2,6 +2,7 @@ November 16, 2021:
 - Add a potential workaround for a graphics engine crash
 - Fix typo in XTARGET_MY_MERCENTARY_TARGET. Use XTARGET_MY_MERCENARY_TARGET instead (#409)
 - Fix Me.Song not returning proper value when Me.Buff also has a match (#411).
+- Add BaseEffectsFocusCap (SongCap) as a member of the Spells TLO
 
 November 10, 2021:
 - Changed the operation of delays in lua so that a delay will no longer block the entire script, only the running thread